requirements.txt
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.3.0
decorator==5.2.1
# diff_gaussian_rasterization @ file:///home/csxq/tasks/SemGauss-SLAM/diff-gaussian-rasterization-w-depth_sem_gauss
# -e git+https://github.com/csxq0605/SemGauss-SLAM.git@77cfca752afcea06614c4edfb03321a39db125e4#egg=dinov2&subdirectory=segmentation/facebookresearch_dinov2_main
# evaluate_3d_reconstruction @ file:///home/csxq/tasks/evaluate_3d_reconstruction_lib
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
Flask==3.1.2
fonttools==4.61.0
fsspec==2025.12.0
fvcore==0.1.5.post20221221
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.1
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
iopath==0.1.10
ipython==8.37.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lightning-utilities==0.15.2
lpips==0.1.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.13.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
plyfile==1.1.3
portalocker==3.2.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
retrying==1.4.2
rich==14.2.0
rich-click==1.9.4
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
termcolor==3.2.0
threadpoolctl==3.6.0
timm==1.0.22
torch==2.9.1
torchmetrics==1.8.2
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.10.1
triton==3.5.1
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
wcwidth==0.2.14
Werkzeug==3.1.4
widgetsnbextension==4.0.15
# xformers==0.0.33+5d4b92a5.d20251029
yacs==0.1.8
zipp==3.23.0
